Bugfix: Fix learning rate rounding error
- Ensured learning rate argument type is double, not integer, in gradDesc declaration in DenseLayer.h

+++ include/DenseLayer.h
@@ -32,7 +32,7 @@ public:
     void backProp(Eigen::MatrixXf nextLayerW, Eigen::MatrixXf nextLayerDz, Eigen::MatrixXf prevLayerA);
     void backProp(Eigen::MatrixXf Y, Eigen::MatrixXf prevLayerA);
 
-    void gradDesc(int learningRate);
+    void gradDesc(double learningRate);
 
     Eigen::MatrixXf getZ();
     Eigen::MatrixXf getA();
